refactor the task3 code and fixing bug
