+++ Training2/thongnq-training2/src/App.js
@@ -1,51 +1,60 @@
-import React, {useState, useRef, useEffect} from "react";
+import React, {useState, useRef} from "react";
 
 import Button from "./task3/Button";
 
 function App() {
 
-  const [userInput, setUserInput] = useState(0);
+  const [count, setCount] = useState(0);
+  const [intervalId, setIntervalId] = useState(0);
 
-  const [error, setError] = useState("");
+  const handleClick = () => {
 
-  const [stopTimer, setStopTimer] = useState(false);
-
-  useEffect(() => {
-    const timer = userInput > 0 && !stopTimer && setInterval(() => setUserInput(userInput - 1), 1000);
-    return () => clearInterval(timer);
-  }, [userInput]);
-
-  const inputRef = useRef(null);
-
-  const handleStartCounter = () => {
     const userInput = inputRef.current.value;
-    if(userInput.trim === ""){
+    if(userInput.trim() === ""){
       setError("Please input a number");
     } else if(isNaN(userInput)){
       setError("Invalid input. Must be a number");
-    } else if (parseInt(Number(userInput)) != userInput){
-      setError("Invalid input. Must be a integer");
     } else if (parseInt(userInput) <= 0){
       setError("Number must be greater than 0");
     } else {
-      setUserInput(parseInt(userInput));
-      setError("");
-      setStopTimer(false);
-    }
+      setCount(parseInt(userInput));
+
+      const newIntervalId = setInterval(() => {
+        console.log(count);
+        setCount(prevCount =>{
+          if(prevCount <= 1 ){
+            clearInterval(newIntervalId);
+            setIntervalId(0);
+            return 0;
+          }
+          return prevCount - 1});
+      }, 1000);
+      setIntervalId(newIntervalId);
+  };
+
+}
+
+  const [error, setError] = useState("");
+
+  const inputRef = useRef(null);
+
+
+  const stopCounter = () => {
+    clearInterval(intervalId);
+        setIntervalId(0);
   }
 
-  const handleStopCounter = () => {
-    setStopTimer(true);
-  }
 
   return (
     <div className="App">
-      {userInput ? <h1>Counter: {userInput}</h1> : null}
+      {count ? <h1>Counter: {count}</h1> : null}
       {error ? <h2>Error: {error}</h2> : null}
       <label for="num">Enter a number</label> <br/>
       <input id="num" type="text" ref={inputRef} /><br/>
-      <Button action={handleStartCounter}>Start</Button>
-      {userInput && !stopTimer ?<Button action={handleStopCounter}>Stop</Button> : null}
+      <Button action={handleClick}>
+        Start
+      </Button>
+      {intervalId ? <Button action={stopCounter}>Stop</Button> : null}
     </div>
   );
 }
